% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximation_singuliere.R
\name{as_approximate_svd}
\alias{as_approximate_svd}
\title{This function approximates a matrix "mat" by an idempotent matrix "result"
using singular value decomposition (SVD)}
\usage{
as_approximate_svd(mat)
}
\arguments{
\item{mat}{the matrix}
}
\value{
the idempotent matrix approximating mat
}
\description{
This function approximates a matrix "mat" by an idempotent matrix "result"
using singular value decomposition (SVD)
}
\examples{
as_approximate_svd(matrix(c(1, 2, 2, 4), nrow = 2))
}
