% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_idempotent.R
\name{as_idempotente}
\alias{as_idempotente}
\title{Cette fonction transforme une matrice en matrice idempotentes}
\usage{
as_idempotente(mat)
}
\arguments{
\item{mat}{la matrice doit avoir les lignes linéairement indépendantes et que le nombre de colonnes soit inférieur ou égale au nombre de ligne #nolint}
}
\value{
la matrice idempotent
}
\description{
Cette fonction transforme une matrice en matrice idempotentes
}
\examples{
as_idempotente(matrix(c(1, 0, 0, 1, 0, 1)))
}
