% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_projection.R
\name{as_projection}
\alias{as_projection}
\title{This function transforms a matrix into a projection matrix}
\usage{
as_projection(mat)
}
\arguments{
\item{mat}{the matrix must have linearly independent rows and the number of columns must be less than or equal to the number of rows #nolint}
}
\value{
the projection matrix
}
\description{
This function transforms a matrix into a projection matrix
}
\examples{
as_projection(matrix(c(1, 0, 0, 1, 0, 1)))
}
