% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_regression_multiple.R
\name{coef_lm}
\alias{coef_lm}
\title{This function calculates the coefficient vector of a multiple linear regression using the least squares method}
\usage{
coef_lm(Y, X)
}
\arguments{
\item{Y}{the vector of the variable to predict}

\item{X}{the matrix of explanatory variables}
}
\value{
a vector of regression coefficients
}
\description{
This function calculates the coefficient vector of a multiple linear regression using the least squares method
}
\examples{
Y <- c(2, 3, 5, 7, 11)  # Variable to predict
X <- matrix(c(1, 1, 1, 1, 1,   # Constant for the intercept
               1, 2, 3, 4, 5), # Explanatory variable
             nrow = 5,
             byrow = FALSE)
coefficients <- coef_lm(Y, X)
}
