% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{cout_error}
\alias{cout_error}
\title{Displays a message as an error, coloring it in red framed by this emoji ❌}
\usage{
cout_error(message)
}
\arguments{
\item{message}{the message to display}
}
\value{
the formatted message
}
\description{
Displays a message as an error, coloring it in red framed by this emoji ❌
}
