% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{cout_info}
\alias{cout_info}
\title{Displays a message as information, coloring it in yellow framed by this emoji ℹ️}
\usage{
cout_info(message)
}
\arguments{
\item{message}{the message to display}
}
\value{
the formatted message
}
\description{
Displays a message as information, coloring it in yellow framed by this emoji ℹ️
}
