% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{cout_success}
\alias{cout_success}
\title{Affiche un message sous forme de succès, en coloriant en rouge encadré par cet émojis ✅}
\usage{
cout_success(message)
}
\arguments{
\item{message}{le message à afficher}
}
\value{
le message formater
}
\description{
Affiche un message sous forme de succès, en coloriant en rouge encadré par cet émojis ✅
}
\examples{
cout_success("Opération réussie")
}
