% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norme_frobenius.R
\name{frob_norm}
\alias{frob_norm}
\title{This function returns the Frobenius norm distance between two matrices
It will be much more used to evaluate the validity of idempotent matrices approximated from an initial matrix}
\usage{
frob_norm(A, B = A * 0)
}
\arguments{
\item{A}{the first matrix}

\item{B}{the second matrix}
}
\value{
returns the Frobenius norm distance
}
\description{
This function returns the Frobenius norm distance between two matrices
It will be much more used to evaluate the validity of idempotent matrices approximated from an initial matrix
}
\examples{
frob_norm(matrix(c(1, 2, 2, 4), nrow = 2), t(matrix(c(1, 2, 2, 4), nrow = 2)))
frob_norm(matrix(c(1, 2, 2, 4), nrow = 2))
}
