% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matrix.R
\name{gen_matrix}
\alias{gen_matrix}
\title{This function generates an idempotent matrix from the given dimensions}
\usage{
gen_matrix(n)
}
\arguments{
\item{n}{defines the dimensions of the matrix n*n}
}
\value{
returns an idempotent matrix of dimensions n*n
}
\description{
This function generates an idempotent matrix from the given dimensions
}
\examples{
gen_matrix(8)
}
