% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_idempotent.R
\name{is_idempotent}
\alias{is_idempotent}
\title{Vérifie si une matrice est idempotente}
\usage{
is_idempotent(mat)
}
\arguments{
\item{mat}{Une matrice carrée.}
}
\value{
TRUE si mat est idempotente (A^2 = A), FALSE sinon.
}
\description{
Vérifie si une matrice est idempotente
}
\examples{
is_idempotent(matrix(c(1, 0, 0, 0), 2, 2))
}
