% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_info.R
\name{matrix_info}
\alias{matrix_info}
\title{This function returns a list containing several pieces of information from a matrix
Notably: its rank, its eigenvalues, its trace, and whether it is idempotent or not}
\usage{
matrix_info(mat)
}
\arguments{
\item{mat}{the input matrix}
}
\value{
a list containing the information of the matrix
}
\description{
This function returns a list containing several pieces of information from a matrix
Notably: its rank, its eigenvalues, its trace, and whether it is idempotent or not
}
\examples{
matrix_info(matrix(c(1, 2, 2, 4), nrow = 2))
matrix_info(matrix(1:6, nrow = 2)) # Non-square matrix
matrix_info(diag(1, 3)) # Idempotent matrix
}
