% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{cout_error}
\alias{cout_error}
\title{Affiche un message sous forme d'erreur, en coloriant en rouge encadré par cet émojis ❌}
\usage{
cout_error(message)
}
\arguments{
\item{message}{le message à afficher}
}
\value{
le message formater
}
\description{
Affiche un message sous forme d'erreur, en coloriant en rouge encadré par cet émojis ❌
}
\examples{
cout_error("Une erreur s'est produite")
}
