% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{cout_info}
\alias{cout_info}
\title{Affiche un message sous forme d'info, en coloriant en rouge encadré par cet émojis ℹ️}
\usage{
cout_info(message)
}
\arguments{
\item{message}{le message à afficher}
}
\value{
le message formater
}
\description{
Affiche un message sous forme d'info, en coloriant en rouge encadré par cet émojis ℹ️
}
\examples{
cout_info("Une information importante")
}
