% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{cout_success}
\alias{cout_success}
\title{Displays a message as a success, coloring it in green framed by this emoji ✅}
\usage{
cout_success(message)
}
\arguments{
\item{message}{the message to display}
}
\value{
the formatted message
}
\description{
Displays a message as a success, coloring it in green framed by this emoji ✅
}
